pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "forgemind-ai"
version = "2.0.0"
description = "ForgeMind AI Suite — Advanced Manufacturing Intelligence Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "ForgeMind Team"},
]

dependencies = [
    # === Core Framework ===
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # === Database ===
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",

    # === Messaging ===
    "aiokafka>=0.10.0",
    "confluent-kafka>=2.3.0",
    "paho-mqtt>=2.0.0",

    # === ML / Data Science ===
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "statsmodels>=0.14.0",
    "prophet>=1.1.5",
    "torch>=2.2.0",
    "tensorflow>=2.15.0",
    "lifelines>=0.28.0",               # Cox PH / survival analysis
    "dowhy>=0.11",                      # Causal inference
    "pgmpy>=0.1.25",                    # Bayesian networks
    "ortools>=9.8",                     # Constraint satisfaction / MILP

    # === Computer Vision ===
    "ultralytics>=8.1.0",              # YOLOv8
    "opencv-python>=4.9.0",
    "Pillow>=10.2.0",
    "grad-cam>=1.5.0",
    "timm>=0.9.12",                    # EfficientNet

    # === OPC-UA / Industrial ===
    "asyncua>=1.1.0",                  # OPC-UA client
    "pyModbusTCP>=0.2.0",

    # === Observability ===
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-prometheus>=0.43b0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "prometheus-client>=0.20.0",

    # === Auth & Security ===
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",

    # === Caching & Utils ===
    "redis>=5.0.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.3",
    "apscheduler>=3.10.4",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "orjson>=3.9.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "A", "COM", "C4", "ISC", "ICN", "PIE", "PT", "RSE", "SIM", "TID", "ARG", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401", "S101", "COM812", "ISC001"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
